requirements.txt
asgiref==3.5.0
better-profanity==0.7.0
boto3==1.26.54
botocore==1.29.165
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.52.0
django-countries==7.5
django-crispy-forms==1.14.0
django-dotenv==1.4.0
django-localflavor==4.0
django-phonenumbers==1.0.1
django-storages==1.13.2
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
phonenumbers==8.13.23
Pillow==9.4.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-stdnum==1.19
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.8.0
tzdata==2021.5
urllib3==1.26.18